+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Send authentication token when retrieving group members.
+
 ## [5.8.0] - 2026-07-31
 
 ### Added